pyproject.toml
[tool.poetry]
name = "percent"
version = "0.1.0"
description = ""
authors = ["Karim Moradi <kmoradi.ai@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.96.0"
uvicorn = "^0.22.0"
matplotlib = "^3.7.1"
jinja2 = "^3.1.2"
databases = {extras = ["postgresql"], version = "^0.7.0"}
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
psycopg2-binary = "^2.9.6"
email-validator = "^2.0.0.post2"
python-dotenv = "^1.0.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.6"
pydantic = {extras = ["email"], version = "^1.10.9"}
alembic = "^1.11.1"
aiofiles = "^23.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
